refactor: er.auth.service , common.auth.service 분리
