+++ src/auth/auth.module.ts
@@ -2,8 +2,9 @@ import { JWT_OPTIONS } from '@config/constant';
 import { jwtOption } from '@config/option';
 import { Global, Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
-import { AuthController } from './controller/er.auth.controller';
-import { AuthService } from './provider/ems.auth.service';
+import { ErAuthController } from './controller/er.auth.controller';
+import { AuthService } from './provider/common.auth.service';
+import { ErAuthService } from './provider/er.auth.service';
 import { EmsJwtAccessStrategy } from './strategy/ems.jwt.access.strategy';
 import { EmsJwtRefreshStrategy } from './strategy/ems.jwt.refresh.strategy';
 import { ErJwtAccessStrategy } from './strategy/er.jwt.access.strategy';
@@ -13,6 +14,8 @@ import { ErJwtRefreshStrategy } from './strategy/er.jwt.refresh.strategy';
   imports: [JwtModule],
   providers: [
     AuthService,
+
+    ErAuthService,
     {
       provide: JWT_OPTIONS,
       useValue: jwtOption,
@@ -22,7 +25,7 @@ import { ErJwtRefreshStrategy } from './strategy/er.jwt.refresh.strategy';
     EmsJwtAccessStrategy,
     EmsJwtRefreshStrategy,
   ],
-  controllers: [AuthController],
-  exports: [AuthService, ErJwtAccessStrategy, EmsJwtAccessStrategy],
+  controllers: [ErAuthController],
+  exports: [ErAuthService, ErJwtAccessStrategy, EmsJwtAccessStrategy, AuthService],
 })
 export class AuthModule {}
