+++ src/auth/controller/ems.auth.controller.ts
@@ -1,0 +1,15 @@
+import { TypedRoute } from '@nestia/core';
+import { Controller } from '@nestjs/common';
+
+@Controller('/ems/auth')
+export class EmsAuthController {
+  @TypedRoute.Get('/')
+  async checkAuthStatus(): Promise<any> {
+    return 'hello';
+  }
+
+  @TypedRoute.Post('/login')
+  async login(): Promise<any> {
+    return 'hello';
+  }
+}
