+++ src/auth/provider/er.auth.service.ts
@@ -1,23 +1,17 @@
 import { PrismaService } from '@common/prisma/prisma.service';
-import { JWT_OPTIONS } from '@config/constant';
 import { AUTH_ERROR } from '@config/errors/auth.error';
-import { JwtOption } from '@config/option/interface';
-import { Inject, Injectable } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { JwtService } from '@nestjs/jwt';
+import { Injectable } from '@nestjs/common';
 import { ErAuthRequest } from '@src/types';
-import * as bcrypt from 'bcrypt';
 import typia from 'typia';
-import { Auth, EmsAuth, ErAuth } from '../interface';
+import { assertPrune } from 'typia/lib/misc';
+import { ErAuth } from '../interface';
+import { AuthService } from './common.auth.service';
 
 @Injectable()
 export class ErAuthService {
   constructor(
     private readonly prismaService: PrismaService,
-    private readonly jwtService: JwtService,
-    private readonly configService: ConfigService,
-    @Inject(JWT_OPTIONS)
-    private readonly jwtOption: JwtOption,
+    private readonly authService: AuthService,
   ) {}
 
   async login({
@@ -43,14 +37,16 @@ export class ErAuthService {
       return typia.random<AUTH_ERROR.EMPLOYEE_NOT_FOUND>();
     }
     const { employee_id, role, hospital_id } = existEmployee;
-    const comparePassword = await this.comparePassword({
+    const comparePassword = await this.authService.comparePassword({
       password,
       hash: existEmployee.password,
     });
     if (!comparePassword) {
       return typia.random<AUTH_ERROR.PASSWORD_INCORRECT>();
     }
-    const { access_token, refresh_token } = this.tokenSign({ emergency_center_id, ...existEmployee });
+    const { access_token, refresh_token } = this.authService.tokenSign(
+      assertPrune<ErAuth.AccessTokenSignPayload>({ emergency_center_id, ...existEmployee }),
+    );
     return {
       access_token,
       refresh_token,
@@ -62,67 +58,5 @@ export class ErAuthService {
         role,
       },
     };
-  }
-
-  async logout() {}
-
-  async hashPassword({ password }: Auth.HashPassword) {
-    const hashedPassword = bcrypt.hash(password, Number(this.configService.get('HASH_SALT')));
-    return hashedPassword;
-  }
-
-  async comparePassword({ password, hash }: Auth.ComparePassword) {
-    return await bcrypt.compare(password, hash);
-  }
-
-  accessTokenSign(accessTokenPayload: ErAuth.AccessTokenSignPayload | EmsAuth.AccessTokenSignPayload) {
-    const access_token = this.jwtService.sign(accessTokenPayload, {
-      secret: this.jwtOption.access_secret,
-      expiresIn: this.jwtOption.access_expires_in,
-    });
-
-    return access_token;
-  }
-
-  accessTokenVerify({ access_token }: Auth.AccessTokenVerify) {
-    try {
-      const verify = this.jwtService.verify<ErAuth.AccessTokenSignPayload | EmsAuth.AccessTokenSignPayload>(
-        access_token,
-        {
-          secret: this.jwtOption.access_secret,
-        },
-      );
-      return verify;
-    } catch (error) {
-      return error;
-    }
-  }
-
-  refreshTokenSign(payload: ErAuth.RefreshTokenSignPayload | EmsAuth.RefreshTokenSignPayload) {
-    const refresh_token = this.jwtService.sign(payload, {
-      secret: this.jwtOption.refresh_secret,
-      expiresIn: this.jwtOption.refresh_expires_in,
-    });
-    return refresh_token;
-  }
-
-  refreshTokenVerify({ refresh_token }: Auth.RefreshTokenVerify) {
-    try {
-      const verify = this.jwtService.verify<ErAuth.RefreshTokenSignPayload | EmsAuth.RefreshTokenSignPayload>(
-        refresh_token,
-        {
-          secret: this.jwtOption.refresh_secret,
-        },
-      );
-      return verify;
-    } catch (error) {
-      return typia.random<AUTH_ERROR.REFRESH_TOKEN_INVALID>();
-    }
-  }
-
-  tokenSign(payload: ErAuth.AccessTokenSignPayload | EmsAuth.AccessTokenSignPayload) {
-    const access_token = this.accessTokenSign(payload);
-    const refresh_token = this.refreshTokenSign(payload);
-    return { access_token, refresh_token };
   }
 }
