+++ src/modules/ems.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { AuthModule } from '@src/auth/auth.module';
+
+@Module({
+  imports: [AuthModule],
+})
+export class EmsModule {}
