+++ tsconfig.json
@@ -116,6 +116,6 @@
       { "transform": "@nestia/core/lib/transform" }
     ]
   },
-  "include": ["src/**/*.ts", ".eslintrc.js"],
+  "include": ["src/**/*.ts", ".eslintrc.js", "src/auth/provider/ems.auth.service.ts"],
   "exclude": ["node_modules", "packages", "src/test/e2e/**/*.ts"]
 }
